+++ zk/create_resp_test.go
@@ -10,7 +10,7 @@ func TestDecodeCreateResp(t *testing.T) {
 	bytes := hex2Bytes(t, "00000001000000000000000500000000000000082f7a6b2d74657374")
 	resp, err := DecodeCreateResp(bytes)
 	assert.Nil(t, err)
-	assert.Equal(t, 1, resp.TransactionId)
+	assert.Equal(t, int32(1), resp.TransactionId)
 	assert.Equal(t, int64(5), resp.ZxId)
 	assert.Equal(t, EC_OK, resp.Error)
 	assert.Equal(t, "/zk-test", resp.Path)
@@ -31,7 +31,7 @@ func TestDecodeCreateRespNodeExistsError(t *testing.T) {
 	bytes := hex2Bytes(t, "000000020000000000000020ffffff92")
 	resp, err := DecodeCreateResp(bytes)
 	assert.Nil(t, err)
-	assert.Equal(t, 2, resp.TransactionId)
+	assert.Equal(t, int32(2), resp.TransactionId)
 	assert.Equal(t, int64(32), resp.ZxId)
 	assert.Equal(t, EC_NodeExistsError, resp.Error)
 }
