+++ zk/get_data_req_test.go
@@ -10,7 +10,7 @@ func TestDecodeGetDataReq(t *testing.T) {
 	bytes := hex2Bytes(t, "0000000300000004000000082f7a6b2d7465737400")
 	req, err := DecodeGetDataReq(bytes)
 	assert.Nil(t, err)
-	assert.Equal(t, 3, req.TransactionId)
+	assert.Equal(t, int32(3), req.TransactionId)
 	assert.Equal(t, OP_GET_DATA, req.OpCode)
 	assert.Equal(t, "/zk-test", req.Path)
 	assert.False(t, req.Watch)
