+++ zk/set_data_req_test.go
@@ -10,7 +10,7 @@ func TestDecodeSetDataReq(t *testing.T) {
 	bytes := hex2Bytes(t, "0000000200000005000000082f7a6b2d7465737400000005776f726c64ffffffff")
 	req, err := DecodeSetDataReq(bytes)
 	assert.Nil(t, err)
-	assert.Equal(t, 2, req.TransactionId)
+	assert.Equal(t, int32(2), req.TransactionId)
 	assert.Equal(t, OP_SET_DATA, req.OpCode)
 	assert.Equal(t, "/zk-test", req.Path)
 	assert.Equal(t, []byte("world"), req.Data)
